renamed TextVariant
