+++ apps/journeys-admin/src/components/Editor/ControlPanel/Attributes/blocks/Typography/Align/index.ts
@@ -1,0 +1,1 @@
+export { Align } from './Align'
