--- apps/journeys-admin/src/components/Editor/ControlPanel/Attributes/blocks/Typography/TextAlign/index.ts
@@ -1,1 +1,0 @@
-export { TextAlign } from './TextAlign'
